pyproject.toml
[tool.poetry]
name = "django-security-headers-example"
version = "0.1.0"
description = ""
authors = ["Steven Pate <steven@laac.dev>"]
license = "BSD-3-Clause"

[tool.poetry.dependencies]
python = "^3.7"
django-environ = "^0.4.5"
django-csp = "^3.5"
django-feature-policy = "^3.0.0"
django = "^2.2.6"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
